+++ src/focus_manager/__tests__/txm_focus_manager-test.js
@@ -123,7 +123,7 @@ describe("TXMFocusManager", () => {
         assert.strictEqual(inputAction.mock.callCount(), 0);
     });
 
-    describe.only("test focus mouse events enabled/disabled", () => {
+    describe("test focus mouse events enabled/disabled", () => {
         const fm = new TXMFocusManager();
 
         const selectAction = mock.fn();